# docs/swagger.yaml
definitions:
  RestApi_UnUpset_internal_delivery_dto.ChangePasswordRequest:
    properties:
      new_password:
        description: Новый пароль пользователя (обязательное поле, минимум 5 символов)
        minLength: 5
        type: string
      old_password:
        description: Текущий пароль пользователя (обязательное поле)
        type: string
    required:
    - new_password
    - old_password
    type: object
  RestApi_UnUpset_internal_delivery_dto.ChangeUsernameRequest:
    properties:
      new_username:
        description: Новое имя пользователя (обязательное поле, от 3 до 50 символов)
        maxLength: 50
        minLength: 3
        type: string
    required:
    - new_username
    type: object
...
  RestApi_UnUpset_internal_delivery_dto.LoginRequest:
    properties:
      email:
        description: Email пользователя (обязательное поле)
        type: string
      password:
        description: Пароль пользователя (обязательное поле)
        type: string
    required:
    - email
    - password
    type: object
  RestApi_UnUpset_internal_delivery_dto.RegisterUserRequest:
    properties:
      email:
        description: Email пользователя (обязательное поле)
        type: string
      password:
        description: Пароль пользователя (обязательное поле, минимум 5 символов)
        minLength: 5
        type: string
      username:
        description: Имя пользователя (обязательное поле, от 3 до 100 символов)
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  RestApi_UnUpset_internal_delivery_dto.Response:
    properties:
      data:
        description: Данные ответа (опционально)
      message:
        description: Сообщение (обычно используется для ошибок)
        type: string
      status:
        description: Статус ответа (success/error)
        type: string
    type: object
...
  RestApi_UnUpset_internal_delivery_dto.UserResponse:
    properties:
      created_at:
        description: Дата создания учетной записи
        type: string
      email:
        description: Email пользователя
        type: string
      id:
        description: Идентификатор пользователя
        type: integer
      username:
        description: Имя пользователя
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для приложения UnUpset
  title: UnUpset API
  version: "1.0"
paths:
...
  /api/user:
    get:
      description: Получение списка всех зарегистрированных пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.UserResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      summary: Получение списка всех пользователей
      tags:
      - users
  /api/user/{id}:
    delete:
      description: Удаление пользователя по ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      summary: Удаление пользователя
      tags:
      - users
    get:
      description: Получение информации о конкретном пользователе по ID
      parameters:
      - description: ID пользователя
        format: uint
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      security:
      - ApiKeyAuth: []
      summary: Получение информации о пользователе
      tags:
      - users
  /api/user/password/{id}:
    patch:
      consumes:
      - application/json
      description: Изменение пароля пользователя
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для смены пароля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      summary: Изменение пароля
      tags:
      - users
  /api/user/username/{id}:
    patch:
      consumes:
      - application/json
      description: Изменение имени пользователя
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Новое имя пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.ChangeUsernameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      summary: Изменение имени пользователя
      tags:
      - users
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя по email и паролю
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      summary: Вход пользователя
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создание нового пользователя
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      summary: Регистрация пользователя
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    description: Для тестирования в Swagger UI (на макбуке логин работает только через
      хром, в сафари протестировать не получается)
    in: cookie
    name: mysession
    type: apiKey
swagger: "2.0"